+++ benchmark/modules/run_1yr_tt_benchmark.py
@@ -385,6 +385,7 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                 restrict_cats=restrict_cats,
                 overwrite=True,
                 spcdb_dir=spcdb_dir,
+                n_job=config["options"]["n_cores"]
             )
 
             # --------------------------------------------------------------
@@ -407,6 +408,7 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                     restrict_cats=restrict_cats,
                     overwrite=True,
                     spcdb_dir=spcdb_dir,
+                    n_job=config["options"]["n_cores"]
                 )
 
         # ==================================================================
@@ -453,6 +455,7 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                     benchmark_type=bmk_type,
                     overwrite=True,
                     spcdb_dir=spcdb_dir,
+                    n_job=config["options"]["n_cores"]
                 )
 
                 # ----------------------------------------------------------
@@ -474,6 +477,7 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                         benchmark_type=bmk_type,
                         overwrite=True,
                         spcdb_dir=spcdb_dir,
+                        n_job=config["options"]["n_cores"]
                     )
 
 
@@ -528,11 +532,16 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                     spcdb_dir=spcdb_dir,
                 )
 
-            # Run in parallel
-            results = Parallel(n_jobs=-1)(
-                delayed(gcc_vs_gcc_mass_table)(mon) \
-                for mon in range(bmk_n_months)
-            )
+            # Create tables in parallel
+            # Turn off parallelization if n_jobs==1
+            if n_jobs != 1:
+                results = Parallel(n_jobs=config["options"]["n_cores"])(
+                    delayed(gcc_vs_gcc_mass_table)(mon) \
+                    for mon in range(bmk_n_months)
+                )
+            else:
+                for mon in range(bmk_n_months):
+                    results = gcc_vs_gcc_mass_table(mon)
 
         # ==================================================================
         # GCC vs GCC operations budgets tables
@@ -666,6 +675,7 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                 restrict_cats=restrict_cats,
                 overwrite=True,
                 spcdb_dir=spcdb_dir,
+                n_job=config["options"]["n_cores"]
             )
 
             # --------------------------------------------------------------
@@ -688,6 +698,7 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                     restrict_cats=restrict_cats,
                     overwrite=True,
                     spcdb_dir=spcdb_dir,
+                    n_job=config["options"]["n_cores"]
                 )
 
         # ==================================================================
@@ -735,6 +746,7 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                     benchmark_type=bmk_type,
                     normalize_by_area=True,
                     spcdb_dir=spcdb_dir,
+                    n_job=config["options"]["n_cores"]
                 )
 
                 # ----------------------------------------------------------
@@ -757,6 +769,7 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                         benchmark_type=bmk_type,
                         normalize_by_area=True,
                         spcdb_dir=spcdb_dir,
+                        n_job=config["options"]["n_cores"]
                     )
 
         # ==================================================================
@@ -815,6 +828,17 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                     gchp_is_pre_14_0=config["data"]["dev"]["gchp"][
                         "is_pre_14.0"]
                 )
+
+            # Create tables in parallel
+            # Turn off parallelization if n_jobs==1
+            if n_jobs != 1:
+                results = Parallel(n_jobs=config["options"]["n_cores"])(
+                    delayed(gchp_vs_gcc_mass_table)(mon) \
+                    for mon in range(bmk_n_months)
+                )
+            else:
+                for mon in range(bmk_n_months):
+                    results = gchp_vs_gcc_mass_table(mon)
 
         # ==================================================================
         # GCHP vs GCC operations budgets tables
@@ -927,7 +951,8 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                 restrict_cats=restrict_cats,
                 overwrite=True,
                 spcdb_dir=spcdb_dir,
-                cmpres=cmpres
+                cmpres=cmpres,
+                n_job=config["options"]["n_cores"]
             )
 
             # --------------------------------------------------------------
@@ -950,7 +975,8 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                     restrict_cats=restrict_cats,
                     overwrite=True,
                     spcdb_dir=spcdb_dir,
-                    cmpres=cmpres
+                    cmpres=cmpres,
+                    n_job=config["options"]["n_cores"]
                 )
 
         # ==================================================================
@@ -1000,7 +1026,8 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                     benchmark_type=bmk_type,
                     normalize_by_area=True,
                     spcdb_dir=spcdb_dir,
-                    cmpres=cmpres
+                    cmpres=cmpres,
+                    n_job=config["options"]["n_cores"]
                 )
 
                 # ----------------------------------------------------------
@@ -1023,7 +1050,8 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                         benchmark_type=bmk_type,
                         normalize_by_area=True,
                         spcdb_dir=spcdb_dir,
-                        cmpres=cmpres
+                        cmpres=cmpres,
+                        n_job=config["options"]["n_cores"]
                     )
 
         # ==================================================================
@@ -1113,11 +1141,16 @@ def run_benchmark(config, bmk_year_ref, bmk_year_dev):
                     dev_met_extra=devareapath
                 )
 
-            # Run in parallel
-            results = Parallel(n_jobs=-1)(
-                delayed(gchp_vs_gchp_mass_table)(mon) \
-                for mon in range(bmk_n_months)
-            )
+            # Create tables in parallel
+            # Turn off parallelization if n_jobs==1
+            if n_jobs != 1:
+                results = Parallel(n_jobs=config["options"]["n_cores"])(
+                    delayed(gchp_vs_gchp_mass_table)(mon) \
+                    for mon in range(bmk_n_months)
+                )
+            else:
+                for mon in range(bmk_n_months):
+                    results = gchp_vs_gchp_mass_table(mon)
 
         # ==================================================================
         # GCHP vs GCHP operations budgets tables